Add QT_NO_NARROWING_CONVERSIONS_IN_CONNECT

+++ source/app/ui/graphquickitem.cpp
@@ -214,7 +214,7 @@ void GraphQuickItem::onLayoutChanged()
     emit layoutChanged();
 }
 
-void GraphQuickItem::onFPSChanged(int fps)
+void GraphQuickItem::onFPSChanged(float fps)
 {
     _fps = fps;
     emit fpsChanged();
